initial commit of metadata fixes:
1. add RecordID to scoring data.
2. only write Score and RecordID back.
3. join with Sim data using RecordID as key.
4. remove RecordID for final output.

One issue with this approach is that the output will contain fields that were not used for scoring. This occurs when the Sim data has fields that were not in the model object, in which case they will not be used for scoring. However, I believe this is an acceptable compromise to the metadata errors that get generated and the added overhead of an AlteryxFullUpdate.